+++ app/assets/javascripts/govuk_publishing_components/dependencies.js
@@ -6,4 +6,13 @@
 
 document.addEventListener('DOMContentLoaded', function () {
   window.GOVUK.modules.start()
+
+  // if statements ensure these functions don't execute during testing
+  if (typeof window.GOVUK.analyticsGa4.vars === 'undefined') {
+    window.GOVUK.loadAnalytics.loadGa4()
+  }
+
+  if (typeof window.GOVUK.analyticsInit === 'undefined') {
+    window.GOVUK.loadAnalytics.loadUa()
+  }
 })