+++ spec/javascripts/helpers/SpecHelper.js
@@ -22,13 +22,19 @@ var resetCookies = function () {
   document.cookie.split(';').forEach(function (c) { document.cookie = c.replace(/^ +/, '').replace(/=.*/, '=;expires=' + new Date().toUTCString() + ';path=/') })
 }
 
+var savedUaVars
+
 // external-link-tracker.js adds a click event to the body for matching links
 // gets called once in production but multiple times in testing e.g. in static-analytics-spec
 // every time it does `new GOVUK.StaticAnalytics()` so need to remove this using .off()
 beforeEach(function () {
   $('body').off()
+
+// load-analytics.js modifies the universal analytics vars, so we need to ensure they are reset each time.
+  savedUaVars = window.GOVUK.extendObject(window.GOVUK.analyticsVars)
 })
 
 afterEach(function () {
   resetCookies()
+  window.GOVUK.analyticsVars = savedUaVars
 })